+++ demo/src/objects/Simple.ts
@@ -1,10 +1,10 @@
-import { CanvasObject, ObjectState, ImageLoader } from 'react-canvas-animate'
+import { CanvasObject, ImageLoader } from 'react-canvas-animate'
 
-export interface SimpleState extends ObjectState {
+export interface SimpleState extends CanvasObject.State {
   images: ImageLoader
 }
 
-export class Simple extends CanvasObject<SimpleState> {
+export class Simple extends CanvasObject.Base<SimpleState> {
   update() {
     // console.log('Simple update', this.state)
   }